Int -> long

+++ src/main/java/com/pehulja/threading/IncrementalJob.java
@@ -4,15 +4,15 @@ import com.pehulja.threading.counters.Counter;
 
 class IncrementalJob implements Runnable {
 	private final Counter counter;
-	private final int iterations;
-	public IncrementalJob(Counter counter, int iterations) {
+	private final long iterations;
+	public IncrementalJob(Counter counter, long iterations) {
 		this.counter = counter;
 		this.iterations = iterations;
 	}
 
 	@Override
 	public void run() {
-		for(int i = 0; i < this.iterations; i++) {
+		for(long i = 0; i < this.iterations; i++) {
 			this.counter.incrementAndGet();
 		}
 	}
