+++ src/main/java/com/pehulja/threading/counters/Counter.java
@@ -1,6 +1,6 @@
 package com.pehulja.threading.counters;
 
 public interface Counter{
-	public int incrementAndGet();
-	public int getValue();
+	public long incrementAndGet();
+	public long getValue();
 }
