+++ src/main/java/com/pehulja/threading/counters/plain/AbstractPlainCounter.java
@@ -3,10 +3,10 @@ package com.pehulja.threading.counters.plain;
 import com.pehulja.threading.counters.Counter;
 
 public abstract class AbstractPlainCounter implements Counter {
-	protected int counter = 0;
+	protected long counter = 0;
 
 	@Override
-	public int getValue() {
+	public long getValue() {
 		return counter;
 	}
 }
