+++ src/main/java/com/pehulja/threading/counters/plain/SynchronizedFieldPlainCounter.java
@@ -2,8 +2,8 @@ package com.pehulja.threading.counters.plain;
 
 public class SynchronizedFieldPlainCounter extends AbstractPlainCounter{
 	@Override
-	public int incrementAndGet() {
-		int value = 0;
+	public long incrementAndGet() {
+		long value = 0;
 		synchronized (this) {
 			value = ++counter;
 		}
